agentforce2/pyproject.toml
[project]
name = "agentforce2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "gradio>=6.12.0",
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]
